Allow to disable sslverify on images download
The default value is still true but in a CI context because we are
using a self-signed certificate, we need to disable the SSL
verification.
This can be done by changing the value of dci_rhel_agent_sslverify
in the settings.yml file to no (or as an extra ansible variable -e).

Change-Id: I6a55f74e7feed23557d620cb1393b660c0ab04a6