Download the .treeinfo file and add image checksum
The .treeinfo file contains the image checksums that can be used to
avoid downloading images on every run even if the images did not
changed.

So we don't need to use the force option with get_url on the images
and we can use the checksum parameter to calculate the digest before
the file download and not after.

With this change the task takes less than 10s compared to several
minutes or more (up to one hour in some case).

The efiboot.img is removed because it's not used by beaker and not
part of the .treeinfo file.

Depends-On: I1d97eb7401adaf8b919c6f0fe9f2446c96455cce
Change-Id: I44e6a0105d3c502ff11ffc67adbd992ffb31a00f