Update patch set 3
Patch Set 3: Code-Review+2

> We need to review together how to handle the beaker "release" task
I guess the release task is currently fine.
Beaker is doing itself the release of the node during beaker job failures so we only have to take care of the node release post running state.
We only have the tests (post-run state) which already handle that in ansible block/rescue so we should be good.
Someone else ?

Patch-set: 3