import: Reuse the .treeinfo file with beaker
Instead of doing a nakedtree import and specify all parameters and
because we are now downloading the .treeinfo file, beaker will use
the information present in that file for the import (distro name,
version, initrd/kernel).
That way we don't stick with the hardcoded RedHatEnterpriseLinux7.
Also this fixes the distro name/release to work with the
beaker-repo-update command required to get the harness repository.

Change-Id: I07de79088b1482867e71e5896c5a5024cf7ee52e