Use components list
Currently we assume that there's three components and:
 - 0 is the RHEL component.
 - 1 the hwcert component.
 - 2 is the extraos component.

Depending on the RHEL release, the number and order of the
components could be different.

This patch allows to download all the components in the same task
(we iterate on the components list returned by the api) and remove
the hardcoded reference to specific component name.

Change-Id: Ib76fcd6a46d4ade140795df69e4e4b52b7b00009