Use --settings instead of crafting the request manually in dci-downloader cmd
Settings file may become complicated with the introduction of the filters.
To avoid crafting a very complex command for dci-downloader, we can use the settings file directly.

This patch send the settings.yml on the beaker server.
Run the dci-downloader command with this temp settings file.
Then remove the temp settings file.

Change-Id: I4a47f3acad91190c97dfc0725ee8bdecf9a2723a