Fix XML exported by bkr job-result command
The output of the bkr job-results command is a binary raw value.
This patch add a python script that escape the output into a proper xml value.
This script is executed on the jumpbox and print the job-result correctly.

Add a workaround folder to capture the fact that this is a workaround.

Change-Id: I24e0982c0e6baf89f9a376f13e4b0dca78bf5a82